Moves back new tests to existing test classes
