+++ tests/phpunit/tests/test-translated-post.php
@@ -276,4 +276,28 @@ class Translated_Post_Test extends PLL_Translated_Object_UnitTestCase {
 		$this->assertNotSame( $db_infos['table'], $multi_db_infos['table'], 'The table name should be different between blogs.' );
 		$this->assertNotSame( $db_infos['default_alias'], $multi_db_infos['default_alias'], 'The field alias should be different between blogs.' );
 	}
+
+	/**
+	 * Checks that updating a post translations group is done only once when we unlink all translations.
+	 */
+	public function test_should_not_update_translations_group_when_removing_all_translations() {
+		$posts = self::factory()->post->create_translated(
+			array( 'lang' => 'en' ),
+			array( 'lang' => 'fr' )
+		);
+
+		$saved_term_count = did_action( 'saved_post_translations' );
+
+		$terms = wp_get_object_terms( $posts, 'post_translations' );
+		$this->assertCount( 1, $terms );
+
+		// Removes the translations from the group by updating the English post.
+		self::$model->post->save_translations( $posts['en'], array() );
+
+		// Checks we updated translations group only once when removing all the translations.
+		$this->assertSame( 1, did_action( 'saved_post_translations' ) - $saved_term_count );
+
+		$this->assertSame( 0, self::$model->post->get_translation( $posts['en'], 'fr' ) );
+		$this->assertSame( 0, self::$model->post->get_translation( $posts['fr'], 'en' ) );
+	}
 }
